Create .gitattributes
The objective of this hackathon is to develop a predictive model that determines if a customer will submit a vehicle insurance claim in the next three months. This solution will help streamline financial planning at AXA and allow them to better serve their customers by understanding which customers are likely to submit a claim.